backend/requirements.txt
asgiref
Django
daphne
django-cors-headers
djangorestframework
pytz
sqlparse
psycopg2-binary
python-dotenv
